requirements.txt

yacs==0.1.8
opencv-python==4.5.1.48
cython
cython-bbox==0.1.3
scipy==1.5.4
numba==0.52.0
progress==1.5
motmetrics==1.2.0
matplotlib==3.3.3
lap==0.4.0
openpyxl==3.0.6
Pillow==8.1.0
tensorboardX==2.1
ipykernel
cord-client-python
moviepy